Create README.md
TGC INDIA offers the best graphic design course in Delhi NCR &amp; UI UX Course in Delhi. Total Graphics Classes (TGC) - is a sincere attempt on part of a few committed professionals in the field of Animation &amp; Digital Art Media. We endeavor to turn out self-driven, top-class professionals in this field who derive the utmost satisfaction from challenges at work and deadlines. This industry focus sets us apart from others in this field and if, we may say so, is also one of our competitive advantages. TGC, an ISO 9001:2008 company, was incorporated as a digital design institute in 2000. We are a favorite among the students as a class design &amp; UI UX course in Delhi India